Add abacusutils to package list
Hi @adematti, we've reduced the primary dependency footprint in https://github.com/abacusorg/abacusutils/pull/46, so I think abacusutils is ready to be tested in cosmodesiconda.  I haven't released a new version on pip with the changes, so I haven't added a tag/version number to `pkg_list.txt` yet.  Are we able to test it like this, and then we can add a version number once we're finished testing and I make the release?